+++ sdk/src/types/block/address/mod.rs
@@ -2,19 +2,17 @@
 // SPDX-License-Identifier: Apache-2.0
 
 mod alias;
-mod bech;
+pub(crate) mod bech32;
 mod ed25519;
 mod nft;
 
 use alloc::{string::String, vec::Vec};
 
-use bech32::{self, FromBase32, ToBase32, Variant};
+use ::bech32::{FromBase32, ToBase32, Variant};
 use derive_more::From;
 use packable::PackableExt;
 
-/// Custom de/serialization for [`address::Bech32`]
-pub(crate) mod bech_serde;
-pub use self::{alias::AliasAddress, bech::Bech32, ed25519::Ed25519Address, nft::NftAddress};
+pub use self::{alias::AliasAddress, bech32::Bech32Address, ed25519::Ed25519Address, nft::NftAddress};
 use crate::types::block::{
     output::{Output, OutputId},
     semantic::{ConflictReason, ValidationContext},
@@ -116,7 +114,7 @@ impl Address {
 
     /// Tries to create an [`Address`] from a bech32 encoded string, also returns the HRP.
     pub fn try_from_bech32_with_hrp<T: AsRef<str>>(address: T) -> Result<(Self, String), Error> {
-        match bech32::decode(address.as_ref()) {
+        match ::bech32::decode(address.as_ref()) {
             Ok((hrp, data, _)) => {
                 let bytes = Vec::<u8>::from_base32(&data).map_err(|_| Error::InvalidAddress)?;
                 Self::unpack_verified(bytes.as_slice(), &())
@@ -130,7 +128,7 @@ impl Address {
     /// Encodes this address to a bech32 string with the given Human Readable Part as prefix.
     pub fn to_bech32<T: AsRef<str>>(&self, hrp: T) -> String {
         // PANIC: encoding can't fail as `self` has already been validated and built.
-        bech32::encode(hrp.as_ref(), self.pack_to_vec().to_base32(), Variant::Bech32).unwrap()
+        ::bech32::encode(hrp.as_ref(), self.pack_to_vec().to_base32(), Variant::Bech32).unwrap()
     }
 
     /// Checks if an string is a valid bech32 encoded address.